feat: Add GitHub workflow for PyPI package release
This commit introduces a new GitHub workflow for releasing Python packages to PyPI. The workflow includes steps for version bumping, package building, and publishing to PyPI. It also includes steps for creating a new release in the GitHub repository and pushing changes with tags. The workflow is triggered manually with the option to specify the version bump type (major, minor, or patch).