test(conftest): change pytest fixture scope from function to session
This commit modifies the scope of the pytest fixture in the conftest.py file.
The scope has been changed from "function" to "session",
which means the fixture will only be invoked once per session,
instead of once per test function.
This change could potentially improve the performance of the test suite
by reducing the number of times the fixture is set up and torn down.